Merged VTK-update-dev-0 into master. All these changes to how the SS deals with VTK are win32 specific. In retrospect, I should have named the branch "VTK-win32-update-*", but what the hell.
Y'all will need to get the latest Windows Prereq installer from the Rat and install it for this to work properly. I have a few more things to do to the new installer, but will send out an email saying it's ready.

Linux behavior should be unchanged from before this merge.